+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## [0.2.42] - 2018-03-25
+### Added
+- The `zip_file()` function.
+- The `allow downloading` modifier.
+- The `centered` feature.
+- The `right` question part.
+### Changed
+- The `allow emailing` modifier now accepts Python code.
+- Margins in navigation bar.
+### Fixed
+- Version of `Flask-User`.
+
 ## [0.2.41] - 2018-03-24
 ### Added
 - The `DACloudStorage` and `DAGoogleAPI` objects.